Main

+++ src/controllers/videoController.js
@@ -1,14 +1,15 @@
-import Video, { formatHashtags } from '../models/Video'
+import Video from '../models/Video'
+import Comment from '../models/Comment'
+import User from '../models/User'
 
 export const home = async (req, res) => {
 	const videos = await Video.find({}).sort({ createdAt: 'desc' })
-	console.log(videos)
 	return res.render('home', { pageTitle: 'Home', videos })
 }
 
 export const watch = async (req, res) => {
 	const { id } = req.params
-	const video = await Video.findById(id)
+	const video = await Video.findById(id).populate('owner').populate('comments')
 	if (!video) {
 		return res.status(404).render('404', { pageTitle: 'Uh-Oh.. Video not found.' })
 	}
@@ -17,25 +18,40 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
 	const { id } = req.params
+	const {
+		user: { _id }
+	} = req.session
 	const video = await Video.findById(id)
 	if (!video) {
 		return res.status(404).render('edit', { pageTitle: 'Uh-Oh.. Video not found.' })
+	}
+	if (String(video.owner) !== String(_id)) {
+		req.flash('error', 'Not authorized')
+		return res.status(403).redirect('/')
 	}
 	return res.render('edit', { pageTitle: `Edit : ${video.title}`, video })
 }
 
 export const postEdit = async (req, res) => {
+	const {
+		user: { _id }
+	} = req.session
 	const { id } = req.params
 	const { title, description, hashtags } = req.body
 	const video = await Video.exists({ _id: id })
 	if (!video) {
 		return res.render('edit', { pageTitle: 'Uh-Oh.. Video not found.' })
 	}
+	if (String(video.owner) !== String(_id)) {
+		req.flash('error', 'You are not the the owner of the video.')
+		return res.status(403).redirect('/')
+	}
 	await Video.findByIdAndUpdate(id, {
 		title,
 		description,
 		hashtags: Video.formatHashtags(hashtags)
 	})
+	req.flash('success', 'Changes saved.')
 	return res.redirect(`/videos/${id}`)
 }
 
@@ -44,13 +60,24 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
+	const {
+		user: { _id }
+	} = req.session
+	const { video, thumb } = req.files
 	const { title, description, hashtags } = req.body
+	const isHeroku = process.env.NODE_ENV === 'production'
 	try {
-		await Video.create({
+		const newVideo = await Video.create({
 			title,
 			description,
+			fileUrl: isHeroku ? video[0].location : video[0].path,
+			thumbUrl: isHeroku ? thumb[0].location : video[0].path,
+			owner: _id,
 			hashtags: Video.formatHashtags(hashtags)
 		})
+		const user = await User.findById(_id)
+		user.videos.push(newVideo._id)
+		user.save()
 		return res.redirect('/')
 	} catch (error) {
 		console.log(error)
@@ -63,6 +90,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
 	const { id } = req.params
+	const {
+		user: { _id }
+	} = req.session
+	const video = await Video.findById(id)
+	if (!video) {
+		return res.status(404).render('404', { pageTitle: 'Uh-Oh.. Video not found.' })
+	}
+	if (String(video.owner) !== String(_id)) {
+		req.flash('error', 'You are not the the owner of the video.')
+		return res.status(403).redirect('/')
+	}
 	await Video.findByIdAndDelete(id)
 	return res.redirect('/')
 }
@@ -78,4 +116,63 @@ export const search = async (req, res) => {
 		}).sort({ createdAt: 'desc' })
 	}
 	return res.render('search', { pageTitle: 'Search', videos })
+}
+
+export const registerView = async (req, res) => {
+	const { id } = req.params
+	const video = await Video.findById(id)
+	if (!video) {
+		return res.sendStatus(404)
+	}
+	video.meta.views = video.meta.views + 1
+	await video.save()
+	return res.sendStatus(200)
+}
+
+export const createComment = async (req, res) => {
+	const {
+		session: { user },
+		body: { text },
+		params: { id }
+	} = req
+
+	const video = await Video.findById(id)
+
+	if (!video) {
+		return res.sendStatus(404)
+	}
+
+	const comment = await Comment.create({
+		text,
+		owner: user._id,
+		video: id
+	})
+	video.comments.push(comment._id)
+	video.save()
+	return res.status(201).json({ newCommentId: comment._id })
+}
+
+export const deleteComment = async (req, res) => {
+	const {
+		session: {
+			user: { _id }
+		},
+		params: { commentId }
+	} = req
+
+	const comment = await Comment.findById(commentId).populate('owner')
+	const videoId = comment.video
+	if (String(_id) !== String(comment.owner._id)) {
+		return res.sendStatus(404)
+	}
+	const video = await Video.findById(videoId)
+	if (!video) {
+		return res.sendStatus(404)
+	}
+
+	video.comments.splice(video.comments.indexOf(commentId), 1)
+	await video.save()
+	await Comment.findByIdAndDelete(commentId)
+
+	return res.sendStatus(200)
 }
